move astype and shift_months tests to their appropriate spots
